Backend/requirements.txt
fastapi
uvicorn
redis
fastapi-redis-cache
pydantic
python-dotenv
httpx
celery
flower
asyncpg
databases
aiosqlite
motor
python-multipart
itsdangerous
starlette-session
user_agents
qrcode
sockets
ultralytics
bcrypt
mysql.connector
fastcore==1.5.29
aiofiles
opencv-python
matplotlib
moviepy